requirements.txt
Django==1.11.1
djangorestframework==3.6.2
djangorestframework-jwt==1.8.0
django-jsonfield==1.0.1
python-dateutil==2.5.3
Mako==1.0.6
django-mako==0.1.4rc0
qiniu==7.0.8
simplejson==3.10.0
beautifulsoup4==4.5.1
pypinyin==0.21.1
requests==2.11.1
celery==3.1.25
django-celery==3.1.17
redis==2.10.5
django-redis==4.6.0
MySQL-python==1.2.5
gunicorn==19.6.0
supervisor==3.3.1
raven==4.0.2
gevent==1.2.2
lxml==4.2.1
ipython==5.7.0
flake8==3.5.0
oss2
phonenumbers
shaw==0.0.8

# comment scikit and flower for now
# matplotlib==2.2.2
# scipy==1.0.1
# scikit-learn==0.19.1
# flower==0.9.2

